requirements.txt
aiogram==3.*
python-dotenv==1.*
SQLAlchemy==2.*
asyncpg==0.*
alembic==1.*
psycopg[binary]==3.*
pydantic==2.*
